move the Mix.Releases.Plugin to Bootloader module
